Website: address Copilot review (PR #357)
- The stub's TOTP secret is now a real Base32 (RFC 4648) string instead
  of one derived from newId(), which embedded a hyphen and broke
  QR/manual enrollment in demo mode.
- Extract the duplicated modal shell into components/Modal.tsx and reuse
  it in SettingsTab and the two-factor modals, so the dialog markup and
  ARIA live in one place.
